+++ src/scopes/internal/SearchReplyImpl.cpp
@@ -129,6 +129,16 @@ bool SearchReplyImpl::push(unity::scopes::CategorisedResult const& result)
 
 bool SearchReplyImpl::push(unity::scopes::Filters const& filters, unity::scopes::FilterState const& filter_state)
 {
+    // basic consistency check
+    try
+    {
+        internal::FilterBaseImpl::validate_filters(filters);
+    }
+    catch (unity::LogicException const &e)
+    {
+        throw unity::LogicException("SearchReplyImpl::push(): Failed to validate filters");
+    }
+
     VariantMap var;
     var["filters"] = internal::FilterBaseImpl::serialize_filters(filters);
     var["filter_state"] = filter_state.serialize();