+++ test/gtest/scopes/OptionSelectorFilter/OptionSelectorFilter_test.cpp
@@ -62,10 +62,12 @@ TEST(OptionSelectorFilter, single_selection)
 
     FilterState fstate;
     EXPECT_FALSE(fstate.has_filter("f1"));
+    EXPECT_FALSE(filter1->has_active_option(fstate));
 
     // enable option1
     filter1->update_state(fstate, option1, true);
     EXPECT_TRUE(fstate.has_filter("f1"));
+    EXPECT_TRUE(filter1->has_active_option(fstate));
     auto active = filter1->active_options(fstate);
     EXPECT_EQ(1u, active.size());
     EXPECT_TRUE(active.find(option1) != active.end());
@@ -95,6 +97,7 @@ TEST(OptionSelectorFilter, multi_selection)
     filter1->update_state(fstate, option1, true);
     filter1->update_state(fstate, option2, true);
     EXPECT_TRUE(fstate.has_filter("f1"));
+    EXPECT_TRUE(filter1->has_active_option(fstate));
     auto active = filter1->active_options(fstate);
     EXPECT_EQ(2u, active.size());
     EXPECT_TRUE(active.find(option1) != active.end());